Code clean up.
Moved test classes defined by the
`TestThatObjectsCanBeEncodeAsJsonAndDecodedViaAJsonDecoder.php`
integration test into their own respective files.
