+++ tests/integration/TestThatObjectsCanBeEncodeAsJsonAndDecodedViaAJsonDecoder.php
@@ -1,13 +1,15 @@
 <?php
 
-include('/home/darling/Git/PHPJsonUtilities/vendor/autoload.php');
+include(
+    str_replace(
+        'tests' . DIRECTORY_SEPARATOR . 'integration',
+        '',
+        __DIR__
+    ) .'vendor/autoload.php'
+);
 
 use \Darling\PHPJsonUtilities\classes\decoders\JsonDecoder;
 use \Darling\PHPJsonUtilities\classes\encoded\data\Json;
-use \Darling\PHPMockingUtilities\classes\mock\values\MockClassInstance;
-use \Darling\PHPReflectionUtilities\classes\utilities\ObjectReflection;
-use \Darling\PHPReflectionUtilities\classes\utilities\Reflection;
-use \Darling\PHPReflectionUtilities\interfaces\utilities\Reflection as ReflectionInterface;
 use \Darling\PHPTextTypes\classes\strings\AlphanumericText;
 use \Darling\PHPTextTypes\classes\strings\ClassString;
 use \Darling\PHPTextTypes\classes\strings\Id;
@@ -16,104 +18,9 @@ use \Darling\PHPTextTypes\classes\strings\SafeText;
 use \Darling\PHPTextTypes\classes\strings\Text;
 use \Darling\PHPTextTypes\classes\strings\UnknownClass;
 use \Darling\PHPUnitTestUtilities\Tests\dev\mock\classes\PrivateMethods;
-
-final class TestClassA
-{
-
-    private mixed $uninitialized;
-
-    public function __construct(private Id $id, private Name $name) {}
-
-
-    public function id(): Id
-    {
-        $this->uninitialized = $this->id;
-        return $this->id;
-    }
-
-    public function name(): Name
-    {
-        $this->uninitialized = $this->name;
-        return $this->name;
-    }
-
-    public function uninitialized(): mixed
-    {
-        return $this->uninitialized;
-    }
-
-}
-
-final class TestClassB
-{
-
-    public string $data = '';
-
-    /**
-     * @param array<mixed> $array
-     */
-    public function __construct(
-        private array $array = [],
-        private bool $bool = false,
-        private float $float = 1.2345,
-        private int $int = 12345,
-        private string $string = '',
-    )
-    {
-        $this->data = strval(
-            json_encode(
-                [
-                    $this->array,
-                    $this->bool,
-                    $this->float,
-                    $this->int,
-                    $this->string
-                ]
-            )
-        );
-    }
-}
-
-/**
- * @template T
- * @implements Iterator<string>
- */
-class TestIterator implements Iterator
-{
-
-    /**
-     *
-     * @param int $position
-     * @param array<int, string> $array
-     *
-     */
-    public function __construct(private int $position = 0, private array $array = []) {
-        if(empty($this->array)) {
-            $this->array = array( "foo", "bar", "baz", "bazzer");
-        }
-    }
-
-    public function rewind(): void {
-        $this->position = 0;
-    }
-
-    public function current(): string {
-        return $this->array[$this->position];
-    }
-
-    public function key(): int {
-        return $this->position;
-    }
-
-    public function next(): void {
-        ++$this->position;
-    }
-
-    public function valid(): bool {
-        return isset($this->array[$this->position]);
-    }
-
-}
+use \Darling\PHPJsonUtilities\tests\dev\test\classes\TestClassA;
+use \Darling\PHPJsonUtilities\tests\dev\test\classes\TestClassB;
+use \Darling\PHPJsonUtilities\tests\dev\test\classes\TestIterator;
 
 $originalObjects = [
     new TestClassA(new Id(), new Name(new Text('Name'))),
